A Bunch Of Things
Mostly visual changes with this one:

New Animated Splash Screen, and Icons (with Themed Icons support!).

Fixed Colors of Status / Navigation Bar.

Added new theme for the main screen to be more inline with the colors of the T2 website.

Code Wise:

App now launches faster.

Fixed the double splash screen issue with Android 12+ Devices.

Updated Version in build.gradle to 1.03 and Version Code to 3 to fix a fatal bug that caused some Android devices to refuse to install the app if they had a previous version of the app installed.